requirements.txt
fastapi
uvicorn
# pytorch
transformers
pandas
sqlalchemy
psycopg2-binary
pymongo
python-dotenv
arxiv